requirements.txt
biopython==1.79
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
csvsort==1.6.1
numpy==1.23.4
pandas==1.5.1
pip==22.2.2
prettytable==3.4.1
psutil==5.9.3
Pygments==2.13.0
pysam==0.19.1
python-dateutil==2.8.2
pytz==2022.5
rich==12.6.0
setuptools==63.4.1
shellingham==1.5.0
six==1.16.0
typer==0.6.1
wcwidth==0.2.5
wheel==0.37.1
